requirements.txt
gradio>=4.0.0
numpy>=1.21.0
pandas>=1.3.0
requests>=2.31.0
tqdm>=4.66.1
scikit-learn>=1.0.0
datasets>=2.17.0
mistralai>=0.0.12
python-dotenv>=1.0.0 